Restructured files & response model minor change
Module files are resturctured, and LoginResponseModelOK which previously
used annotation @Data is changed to @Getter only as we do not want to
set its values other than from constructor.
