+++ service/src/main/java/com/oasis/service/LoginService.java
@@ -1,9 +1,7 @@
 package com.oasis.service;
 
-import com.oasis.exception.UserNotFoundException;
 import com.oasis.model.EmployeeModel;
 import com.oasis.repository.LoginRepository;
-import com.oasis.webmodel.response.LoginResponseModelOK;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
