+++ web-controller/src/main/java/com/oasis/MappingValues.java
@@ -1,4 +1,4 @@
-package com.oasis.webcontroller;
+package com.oasis;
 
 public class MappingValues {
     public static final String LOGIN_MAPPING_VALUE = "/login";
