+++ web-controller/src/main/java/com/oasis/configuration/DatabaseConfiguration.java
@@ -10,7 +10,7 @@ import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 
 @Configuration
 @EnableWebMvc
-public class OASISConfiguration {
+public class DatabaseConfiguration {
     @Bean
     public MongoDbFactory mongoDbFactory() {
         MongoClient mongoClient = new MongoClient("localhost", 27017);
